added new features and changed Long to Integer

+++ src/main/java/com/fitness/security/entity/Program.java
@@ -2,8 +2,6 @@ package com.fitness.security.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
-
-import java.util.List;
 @Table(name = "program")
 @Entity
 @Data
@@ -13,7 +11,7 @@ public class Program {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     private String name;
