+++ src/main/java/com/fitness/security/repository/ProgramDayRelationRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProgramDayRelationRepository extends JpaRepository<ProgramDayRelation, Long> {
+public interface ProgramDayRelationRepository extends JpaRepository<ProgramDayRelation, Integer> {
 }
